move dto to own package
