+++ jwt.go
@@ -12,6 +12,7 @@ import (
 	jwt "github.com/golang-jwt/jwt"
 	"github.com/google/uuid"	
 	c "github.com/na50r/wombo-combo-go-be/constants"
+	dto "github.com/na50r/wombo-combo-go-be/dto"
 )
 
 func getToken(r *http.Request) (string, bool) {
@@ -137,24 +138,24 @@ func withAccountAuth(handlerFunc http.HandlerFunc) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		token, tokenExists := getToken(r)
 		if !tokenExists {
-			WriteJSON(w, http.StatusUnauthorized, APIError{Error: c.Unauthorized})
+			WriteJSON(w, http.StatusUnauthorized, dto.APIError{Error: c.Unauthorized})
 			log.Println("Unauthorized (No Token)")
 			return
 		}
 		accountClaims, err := verifyAccountJWT(token)
 		if err != nil {
-			WriteJSON(w, http.StatusUnauthorized, APIError{Error: c.Unauthorized})
+			WriteJSON(w, http.StatusUnauthorized, dto.APIError{Error: c.Unauthorized})
 			log.Println("Unauthorized (Invalid Token)", err)
 			return
 		}
 		username, err := GetUsername(r)
 		if err != nil {
-			WriteJSON(w, http.StatusUnauthorized, APIError{Error: c.Unauthorized})
+			WriteJSON(w, http.StatusUnauthorized, dto.APIError{Error: c.Unauthorized})
 			log.Println("Unauthorized (No Username)", err)
 			return
 		}
 		if accountClaims.Username != username {
-			WriteJSON(w, http.StatusUnauthorized, APIError{Error: c.Unauthorized})
+			WriteJSON(w, http.StatusUnauthorized, dto.APIError{Error: c.Unauthorized})
 			log.Println("Unauthorized (Invalid Username)", err)
 			return
 		}
@@ -169,30 +170,30 @@ func withPlayerAuth(handlerFunc http.HandlerFunc) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		token, tokenExists := getToken(r)
 		if !tokenExists {
-			WriteJSON(w, http.StatusUnauthorized, APIError{Error: c.Unauthorized})
+			WriteJSON(w, http.StatusUnauthorized, dto.APIError{Error: c.Unauthorized})
 			log.Println("Unauthorized (No Token)")
 			return
 		}
 		playerClaims, err := verifyPlayerJWT(token)
 		if err != nil {
-			WriteJSON(w, http.StatusUnauthorized, APIError{Error: c.Unauthorized})
+			WriteJSON(w, http.StatusUnauthorized, dto.APIError{Error: c.Unauthorized})
 			log.Println("Unauthorized (Invalid Token)", err)
 			return
 		}
 		lobbyCode, err := GetLobbyCode(r)
 		if err != nil {
-			WriteJSON(w, http.StatusUnauthorized, APIError{Error: c.Unauthorized})
+			WriteJSON(w, http.StatusUnauthorized, dto.APIError{Error: c.Unauthorized})
 			log.Println("Unauthorized (No Lobby Code)", err)
 			return
 		}
 		playerName, err := GetPlayername(r)
 		if err != nil {
-			WriteJSON(w, http.StatusUnauthorized, APIError{Error: c.Unauthorized})
+			WriteJSON(w, http.StatusUnauthorized, dto.APIError{Error: c.Unauthorized})
 			log.Println("Unauthorized (No Player Name)", err)
 			return
 		}
 		if lobbyCode != playerClaims.LobbyCode || playerName != playerClaims.PlayerName {
-			WriteJSON(w, http.StatusUnauthorized, APIError{Error: c.Unauthorized})
+			WriteJSON(w, http.StatusUnauthorized, dto.APIError{Error: c.Unauthorized})
 			log.Println("Unauthorized (Invalid Lobby Code or Player Name)", err)
 			return
 		}
